backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "anthropic>=0.86.0",
    "duckdb>=1.5.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "openai>=2.29.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
